requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
requests>=2.31.0
feedparser>=6.0.11
python-dotenv>=1.0.1
openai>=1.40.0